airflow/requirements.txt
apache-airflow-providers-google 
selenium==3.141.0
beautifulsoup4
pandas
python-dotenv
gcsfs